cleanup to prepare submission
